+++ src/archetypes/Crowdloan/Countdown.tsx
@@ -4,13 +4,13 @@ import { useEffect, useState } from 'react'
 import styled from 'styled-components'
 
 type OngoingProps = {
+  relayChainId: number
   end?: number
   showSeconds?: boolean
   className?: string
-  relayChainId?: number
 }
 
-const Ongoing: React.FC<OngoingProps> = ({ end, showSeconds, className = '', relayChainId }) => {
+const Ongoing: React.FC<OngoingProps> = ({ end, showSeconds, relayChainId, className = '' }) => {
   const [secondsRemaining, setSecondsRemaining] = useState<number>()
   const blockNumber = useChainmetaValue(relayChainId, 'blockNumber')
   const blockPeriod = useChainmetaValue(relayChainId, 'blockPeriod')
@@ -47,15 +47,15 @@ const Countdown: React.FC<CountdownProps> = ({ id, showSeconds, className, ...re
   const { crowdloan } = useCrowdloanById(id)
 
   // Pendor
-  if (!crowdloan) return <Ongoing relayChainId={crowdloan?.relayChainId} />
+  if (!crowdloan) return <Ongoing />
 
-  const { uiStatus, lockExpiredBlock } = crowdloan
+  const { uiStatus, lockExpiredBlock, relayChainId } = crowdloan
 
   if (['active', 'capped'].includes(uiStatus)) {
-    return <Ongoing {...rest} showSeconds={showSeconds} end={lockExpiredBlock} relayChainId={crowdloan?.relayChainId} />
+    return <Ongoing {...rest} showSeconds={showSeconds} end={lockExpiredBlock} relayChainId={relayChainId}/>
   }
-  if (uiStatus === 'winner') return <Generic text="Winner" relayChainId={crowdloan?.relayChainId} />
-  if (uiStatus === 'ended') return <Generic text="Ended" relayChainId={crowdloan?.relayChainId} />
+  if (uiStatus === 'winner') return <Generic text="Winner" />
+  if (uiStatus === 'ended') return <Generic text="Ended" />
 
   // Pendor
   return <Ongoing />
