+++ src/archetypes/Crowdloan/Indexx.tsx
@@ -1,9 +1,11 @@
-import { Crowdloan } from '@archetypes'
 import { Await, Field, Grid, NoResults } from '@components'
+import NoCrowdloans from '@components/NoCrowdloansPlaceholder'
 import { trackGoal } from '@libs/fathom'
 import { device } from '@util/breakpoints'
 import { useTranslation } from 'react-i18next'
 import styled from 'styled-components'
+import { useFilter } from './_hooks'
+import Teaser from './Teaser'
 
 const FilterBar = styled(
   ({
@@ -27,28 +29,10 @@ const FilterBar = styled(
     const { t } = useTranslation()
     return (
       <div className={`${className} filterbar`} {...rest}>
-        <Field.Search
-          className="searchbar"
-          value={search}
-          placeholder={t('Search Crowdloans')}
-          onChange={(search: any) => {
-            setSearch(search)
-            trackGoal('9XUF7WEB', 1) // crowdloan_search
-          }}
-        />
         <div className="filters">
           <Field.RadioGroup
-            value={status}
-            onChange={(status: any) => {
-              setStatus(status)
-              trackGoal('0AO7IT2G', 1) // crowdloan_filter
-            }}
-            options={statusOptions}
-            small
-            secondary
-          />
-          <Field.RadioGroup
             value={network}
+            className="network-filter"
             onChange={(network: any) => {
               setNetwork(network)
               trackGoal('0AO7IT2G', 1) // crowdloan_filter
@@ -57,7 +41,26 @@ const FilterBar = styled(
             small
             primary
           />
+          <Field.Search
+            className="searchbar"
+            value={search}
+            placeholder={t('Search Crowdloans')}
+            onChange={(search: any) => {
+              setSearch(search)
+              trackGoal('9XUF7WEB', 1) // crowdloan_search
+            }}
+          />
         </div>
+        <Field.RadioGroup
+          value={status}
+          onChange={(status: any) => {
+            setStatus(status)
+            trackGoal('0AO7IT2G', 1) // crowdloan_filter
+          }}
+          options={statusOptions}
+          small
+          secondary
+        />
       </div>
     )
   }
@@ -66,7 +69,21 @@ const FilterBar = styled(
   display: flex;
   align-items: center;
   gap: 2rem;
+  justify-content: space-between;
   flex-wrap: wrap;
+
+  .network-filter {
+    align-items: unset;
+    flex-direction: unset;
+
+    .children {
+      border-radius: 1rem;
+    }
+
+    .children .pill {
+      border-radius: inherit;
+    }
+  }
 
   .searchbar {
     display: inline-block;
@@ -78,40 +95,109 @@ const FilterBar = styled(
 
   .filters {
     display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-    justify-content: space-between;
-    width: 100%;
+    flex-wrap: wrap-reverse;
+    gap: 2rem;
+    // justify-content: space-between;
+    // width: 100%;
   }
 `
 
+// const PopularCrowdloans = styled(({ className }) => {
+//   const { crowdloans } = Crowdloan.useFilter()
+//   return (
+//     <Panel className={className}>
+//       <PanelSection className="card">
+//         <div className="title">Popular Crowdloans</div>
+//         <ol className="list">
+//           {crowdloans.slice(0, 5).map(({ id }) => (
+//             <li key={id}>
+//               <CrowdloanSummary id={id} />
+//             </li>
+//           ))}
+//         </ol>
+//       </PanelSection>
+//     </Panel>
+//   )
+// })`
+//   min-width: max-content;
+
+//   .card {
+//     padding: 2.4rem;
+//   }
+
+//   .title {
+//     color: var(--color-dim);
+//   }
+
+//   .list {
+//     margin-top: 2rem;
+
+//     & > * + * {
+//       margin-top: 1rem;
+//     }
+//   }
+
+//   ol {
+//     list-style: none;
+//     counter-reset: popular-crowdloans-counter;
+//     padding-inline: 0;
+//     margin-block-end: 0;
+//   }
+//   ol li {
+//     counter-increment: popular-crowdloans-counter;
+//     display: flex;
+//     align-items: center;
+//     gap: 1rem;
+//   }
+//   ol li::before {
+//     content: counter(popular-crowdloans-counter);
+//     color: var(--color-primary);
+//     font-weight: bold;
+//     max-width: 1rem;
+//     padding-right: 1rem;
+//   }
+// `
+
+// const LearnCrowdloansBanner = () => {
+//   return (
+//     <Panel>
+//       <PanelSection>LearnCrowdloansBanner</PanelSection>
+//     </Panel>
+//   )
+// }
+
 const Index = styled(({ withFilter, className }) => {
   const { t } = useTranslation()
-  const { crowdloans, count, loading, filterProps } = Crowdloan.useFilter()
-
+  const { crowdloans, count, loading, filterProps } = useFilter()
+  
   return (
     <div className={`crowdloan-index ${className}`}>
-      {/* TODO: Remove for now as no Learn more link yet. */}
-      {/* <UnlockTalismanBanner /> */}
+      <div className="overview">
+        {/* TODO: Comment out for now */}
+        {/* <PopularCrowdloans /> */}
+        {/* <LearnCrowdloansBanner />*/}
+        {/* <UnlockTalismanBanner /> */}
+      </div>
       {withFilter && <FilterBar {...filterProps} count={count} />}
       <Await until={!loading}>
-        <NoResults
-          require={count?.filtered > 0}
-          title={t('noResult.title')}
-          subtitle={t('noResult.subtitle')}
-          text={t('noResult.text')}
-        >
+        <NoCrowdloans require={count?.filtered > 0} text={t('noCrowdloans.text')} subtext={t('noCrowdloans.subtext')}>
           <Grid>
             {crowdloans.map(({ id }) => (
-              <Crowdloan.Teaser key={id} id={id} />
+              <Teaser key={id} id={id} />
             ))}
           </Grid>
-        </NoResults>
+        </NoCrowdloans>
       </Await>
     </div>
   )
 })`
   padding: 2.4rem;
+
+  .overview {
+    display: flex;
+    gap: 2rem;
+    flex-wrap: wrap;
+  }
 
   .await {
     font-size: var(--font-size-xxlarge);
