+++ src/archetypes/Crowdloan/index.tsx
@@ -5,5 +5,4 @@ export { default as Index } from './Indexx'
 export { default as Raised } from './Raised'
 export { default as Rewards } from './Rewards'
 export { default as Teaser } from './Teaser'
-export { default as Bonus } from './Bonus'
 export { useFilter } from './_hooks'
