--- src/archetypes/CrowdloanV2/Contributors.tsx
@@ -1,12 +1,0 @@
-// import { useCrowdloanByParachainId } from '@libs/talisman'
-import styled from 'styled-components'
-
-const Contributors = styled(({ id, className }) => {
-  // const {
-  //   item: { crowdloan },
-  // } = useCrowdloanByParachainId(id)
-
-  return <div className={`crowdloan-contributors ${className}`}>todo</div>
-})``
-
-export default Contributors
