--- src/archetypes/CrowdloanV2/index.tsx
@@ -1,8 +1,0 @@
-export { default as Contribute } from './Contribute'
-export { default as Contributors } from './Contributors'
-export { default as Countdown } from './Countdown'
-export { default as Index } from './Indexx'
-export { default as Raised } from './Raised'
-export { default as Rewards } from './Rewards'
-export { default as Teaser } from './Teaser'
-export { useFilter } from './_hooks'
