+++ src/archetypes/index.tsx
@@ -1,5 +1,4 @@
 export * as Account from './Account'
 export * as Parachain from './Parachain'
 export * as Crowdloan from './Crowdloan'
-export * as CrowdloanV2 from './CrowdloanV2'
 export * as Wallet from './Wallet'
