+++ src/routes/Crowdloan.Index.tsx
@@ -1,4 +1,4 @@
-import { CrowdloanV2 as Crowdloan } from '@archetypes'
+import { Crowdloan } from '@archetypes'
 import styled from 'styled-components'
 
 const CrowdloanIndex = styled(({ className }) => (
